modularized functions modals msg productd duplicated
