+++ src/components/modal/ModalFunctions.js
@@ -1,0 +1,34 @@
+import { useState } from 'react';
+
+//set state class to modal 'Successfully Product Added Msg'
+export const ModalFunctions = () => {
+  const [modalClass, setmodalClass] = useState(false);
+  const showModal = () => {
+    setmodalClass(true);
+  };
+  const hideModal = () => {
+    setmodalClass(false);
+  };
+  return {
+    modalClass,
+    showModal,
+    hideModal,
+  };
+};
+//set state class to modal 'Duplicated Product Msg'
+export const ModalFunctionsMssgDuplicatedProduct = () => {
+  const [msgProductDuplicatedModal, setMssgProductDuplicatedModal] =
+    useState(false);
+  //hidden and show modal 'Duplicated Product Msg' functions
+  const showModalMssgProductDuplicated = () => {
+    setMssgProductDuplicatedModal(true);
+  };
+  const hideModalMssgProductDuplicated = () => {
+    setMssgProductDuplicatedModal(false);
+  };
+  return {
+    msgProductDuplicatedModal,
+    showModalMssgProductDuplicated,
+    hideModalMssgProductDuplicated,
+  };
+};
